+++ aftiktuna-rust/aftiktuna-three-d/src/game/frame_render.rs
@@ -6,7 +6,7 @@ use aftiktuna::core::display::ModelId;
 use aftiktuna::core::position::Direction;
 use aftiktuna::core::store::StoreStock;
 use aftiktuna::view::area::{ObjectRenderData, RenderData, RenderProperties};
-use aftiktuna::view::{self, Frame};
+use aftiktuna::view::{self, DialogueExpression, Frame};
 use aftiktuna_three_d::{render, Camera};
 use rand::{Rng, SeedableRng};
 use three_d::Object;
@@ -64,6 +64,7 @@ pub fn render_frame(
                     is_badly_hurt: data.is_badly_hurt,
                     ..RenderProperties::default()
                 },
+                data.expression,
                 &mut assets.aftik_colors,
                 frame_input.accumulated_time as f32,
                 &frame_input.context,
@@ -133,6 +134,7 @@ fn draw_area_view(
                 assets.models.lookup_model(&object.model_id),
                 *pos,
                 &object.properties,
+                DialogueExpression::default(),
                 &mut assets.aftik_colors,
                 frame_input.accumulated_time as f32 + time_offset,
                 &frame_input.context,
@@ -151,6 +153,7 @@ fn draw_area_view(
                             direction: object.properties.direction,
                             ..RenderProperties::default()
                         },
+                        DialogueExpression::default(),
                         &mut assets.aftik_colors,
                         frame_input.accumulated_time as f32,
                         &frame_input.context,
@@ -281,6 +284,7 @@ fn draw_store_view(
             aftik_color: view.shopkeeper_color.clone(),
             ..RenderProperties::default()
         },
+        DialogueExpression::Neutral,
         &mut assets.aftik_colors,
         frame_input.accumulated_time as f32,
         &frame_input.context,