+++ aftiktuna-rust/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - New creature Blood mantis
 - New indoor decorations
 - Ship dialogue
+- Dialogue expressions
 - Json location definition `crew_ship.json` for the ship and new symbol types "ship_controls" and "food_deposit"
 - Noise message when something happens in a nearby area